requirements.txt
# IMF Bz Forecasting Pipeline Requirements
# Install with: pip install -r requirements.txt

# Core dependencies (should already be installed from main Surya)
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0

# Data handling
xarray>=0.20.0
h5py>=3.0.0
netCDF4>=1.5.0
requests>=2.25.0

# Model training
peft>=0.4.0  # For LoRA
timm>=0.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Optional but recommended
wandb>=0.15.0  # For experiment tracking
jupyter>=1.0.0  # For notebooks
tqdm>=4.60.0  # Progress bars
